+++ apollo-router/src/plugins/telemetry/config_new/connector/events.rs
@@ -123,3 +123,201 @@ impl CustomEvents<ConnectorRequest, ConnectorResponse, (), ConnectorAttributes, 
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::str::FromStr;
+
+    use apollo_compiler::name;
+    use apollo_federation::sources::connect::ConnectId;
+    use apollo_federation::sources::connect::ConnectSpec;
+    use apollo_federation::sources::connect::Connector;
+    use apollo_federation::sources::connect::HttpJsonTransport;
+    use apollo_federation::sources::connect::JSONSelection;
+    use apollo_federation::sources::connect::StringTemplate;
+    use http::HeaderValue;
+    use tracing::instrument::WithSubscriber;
+
+    use super::*;
+    use crate::assert_snapshot_subscriber;
+    use crate::plugins::connectors::handle_responses::MappedResponse;
+    use crate::plugins::connectors::make_requests::ResponseKey;
+    use crate::plugins::telemetry::Telemetry;
+    use crate::plugins::test::PluginTestHarness;
+    use crate::services::connector::request_service::Request;
+    use crate::services::connector::request_service::Response;
+    use crate::services::connector::request_service::TransportRequest;
+    use crate::services::connector::request_service::TransportResponse;
+    use crate::services::connector::request_service::transport;
+    use crate::services::router::body;
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_connector_events_request() {
+        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
+            .config(include_str!("../../testdata/custom_events.router.yaml"))
+            .build()
+            .await
+            .expect("test harness");
+
+        async {
+            let context = crate::Context::default();
+            let mut http_request = http::Request::builder().body("".into()).unwrap();
+            http_request
+                .headers_mut()
+                .insert("x-log-request", HeaderValue::from_static("log"));
+            let transport_request = TransportRequest::Http(transport::http::HttpRequest {
+                inner: http_request,
+                debug: None,
+            });
+            let connector = Connector {
+                id: ConnectId::new(
+                    "subgraph".into(),
+                    Some("source".into()),
+                    name!(Query),
+                    name!(users),
+                    0,
+                    "label",
+                ),
+                transport: HttpJsonTransport {
+                    source_url: None,
+                    connect_template: StringTemplate::from_str("/test").unwrap(),
+                    ..Default::default()
+                },
+                selection: JSONSelection::empty(),
+                config: None,
+                max_requests: None,
+                entity_resolver: None,
+                spec: ConnectSpec::V0_1,
+                request_variables: Default::default(),
+                response_variables: Default::default(),
+                batch_settings: None,
+                request_headers: Default::default(),
+                response_headers: Default::default(),
+            };
+            let response_key = ResponseKey::RootField {
+                name: "hello".to_string(),
+                inputs: Default::default(),
+                selection: Arc::new(JSONSelection::parse("$.data").unwrap()),
+            };
+            let connector_request = Request {
+                context: context.clone(),
+                connector: Arc::new(connector.clone()),
+                service_name: Default::default(),
+                transport_request,
+                key: response_key.clone(),
+                mapping_problems: vec![],
+                supergraph_request: Default::default(),
+            };
+            test_harness
+                .call_connector_request_service(connector_request, |request| Response {
+                    context: request.context.clone(),
+                    connector: request.connector.clone(),
+                    transport_result: Ok(TransportResponse::Http(transport::http::HttpResponse {
+                        inner: http::Response::builder()
+                            .status(200)
+                            .header("x-log-request", HeaderValue::from_static("log"))
+                            .body(body::empty())
+                            .expect("expecting valid response")
+                            .into_parts()
+                            .0,
+                    })),
+                    mapped_response: MappedResponse::Data {
+                        data: serde_json::json!({})
+                            .try_into()
+                            .expect("expecting valid JSON"),
+                        key: request.key.clone(),
+                        problems: vec![],
+                    },
+                })
+                .await
+                .expect("expecting successful response");
+        }
+        .with_subscriber(assert_snapshot_subscriber!())
+        .await
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_connector_events_response() {
+        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
+            .config(include_str!("../../testdata/custom_events.router.yaml"))
+            .build()
+            .await
+            .expect("test harness");
+
+        async {
+            let context = crate::Context::default();
+            let mut http_request = http::Request::builder().body("".into()).unwrap();
+            http_request
+                .headers_mut()
+                .insert("x-log-response", HeaderValue::from_static("log"));
+            let transport_request = TransportRequest::Http(transport::http::HttpRequest {
+                inner: http_request,
+                debug: None,
+            });
+            let connector = Connector {
+                id: ConnectId::new(
+                    "subgraph".into(),
+                    Some("source".into()),
+                    name!(Query),
+                    name!(users),
+                    0,
+                    "label",
+                ),
+                transport: HttpJsonTransport {
+                    source_url: None,
+                    connect_template: StringTemplate::from_str("/test").unwrap(),
+                    ..Default::default()
+                },
+                selection: JSONSelection::empty(),
+                config: None,
+                max_requests: None,
+                entity_resolver: None,
+                spec: ConnectSpec::V0_1,
+                request_variables: Default::default(),
+                response_variables: Default::default(),
+                batch_settings: None,
+                request_headers: Default::default(),
+                response_headers: Default::default(),
+            };
+            let response_key = ResponseKey::RootField {
+                name: "hello".to_string(),
+                inputs: Default::default(),
+                selection: Arc::new(JSONSelection::parse("$.data").unwrap()),
+            };
+            let connector_request = Request {
+                context: context.clone(),
+                connector: Arc::new(connector.clone()),
+                service_name: Default::default(),
+                transport_request,
+                key: response_key.clone(),
+                mapping_problems: vec![],
+                supergraph_request: Default::default(),
+            };
+            test_harness
+                .call_connector_request_service(connector_request, |request| Response {
+                    context: request.context.clone(),
+                    connector: request.connector.clone(),
+                    transport_result: Ok(TransportResponse::Http(transport::http::HttpResponse {
+                        inner: http::Response::builder()
+                            .status(200)
+                            .header("x-log-response", HeaderValue::from_static("log"))
+                            .body(body::empty())
+                            .expect("expecting valid response")
+                            .into_parts()
+                            .0,
+                    })),
+                    mapped_response: MappedResponse::Data {
+                        data: serde_json::json!({})
+                            .try_into()
+                            .expect("expecting valid JSON"),
+                        key: request.key.clone(),
+                        problems: vec![],
+                    },
+                })
+                .await
+                .expect("expecting successful response");
+        }
+        .with_subscriber(assert_snapshot_subscriber!())
+        .await
+    }
+}
