+++ apollo-router/src/plugins/telemetry/config_new/events.rs
@@ -2,12 +2,7 @@ use std::fmt::Debug;
 use std::marker::PhantomData;
 use std::sync::Arc;
 
-#[cfg(test)]
-use http::HeaderValue;
-use opentelemetry::Key;
 use opentelemetry::KeyValue;
-use opentelemetry_semantic_conventions::trace::HTTP_REQUEST_METHOD;
-use parking_lot::Mutex;
 use schemars::JsonSchema;
 use serde::Deserialize;
 use tower::BoxError;
@@ -17,16 +12,10 @@ use tracing::info_span;
 use super::Selector;
 use super::Selectors;
 use super::Stage;
+use super::router::events::RouterEvents;
+use super::subgraph::events::SubgraphEvents;
+use super::supergraph::events::SupergraphEvents;
 use crate::Context;
-use crate::graphql;
-use crate::plugins::telemetry::config_new::attributes::HTTP_REQUEST_BODY;
-use crate::plugins::telemetry::config_new::attributes::HTTP_REQUEST_HEADERS;
-use crate::plugins::telemetry::config_new::attributes::HTTP_REQUEST_URI;
-use crate::plugins::telemetry::config_new::attributes::HTTP_REQUEST_VERSION;
-use crate::plugins::telemetry::config_new::attributes::HTTP_RESPONSE_BODY;
-use crate::plugins::telemetry::config_new::attributes::HTTP_RESPONSE_HEADERS;
-use crate::plugins::telemetry::config_new::attributes::HTTP_RESPONSE_STATUS;
-use crate::plugins::telemetry::config_new::attributes::HTTP_RESPONSE_VERSION;
 use crate::plugins::telemetry::config_new::conditions::Condition;
 use crate::plugins::telemetry::config_new::connector::attributes::ConnectorAttributes;
 use crate::plugins::telemetry::config_new::connector::events::ConnectorEvents;
@@ -34,15 +23,15 @@ use crate::plugins::telemetry::config_new::connector::events::ConnectorEventsCon
 use crate::plugins::telemetry::config_new::connector::selectors::ConnectorSelector;
 use crate::plugins::telemetry::config_new::extendable::Extendable;
 use crate::plugins::telemetry::config_new::router::attributes::RouterAttributes;
+use crate::plugins::telemetry::config_new::router::events::RouterEventsConfig;
 use crate::plugins::telemetry::config_new::selectors::RouterSelector;
 use crate::plugins::telemetry::config_new::selectors::SubgraphSelector;
 use crate::plugins::telemetry::config_new::selectors::SupergraphSelector;
 use crate::plugins::telemetry::config_new::subgraph::attributes::SubgraphAttributes;
+use crate::plugins::telemetry::config_new::subgraph::events::SubgraphEventsConfig;
 use crate::plugins::telemetry::config_new::supergraph::attributes::SupergraphAttributes;
+use crate::plugins::telemetry::config_new::supergraph::events::SupergraphEventsConfig;
 use crate::plugins::telemetry::dynamic_attribute::EventDynAttribute;
-use crate::services::router;
-use crate::services::subgraph;
-use crate::services::supergraph;
 
 #[derive(Clone)]
 pub(crate) struct DisplayRouterRequest(pub(crate) EventLevel);
@@ -176,20 +165,6 @@ impl Events {
     }
 }
 
-pub(crate) type RouterEvents =
-    CustomEvents<router::Request, router::Response, (), RouterAttributes, RouterSelector>;
-
-pub(crate) type SupergraphEvents = CustomEvents<
-    supergraph::Request,
-    supergraph::Response,
-    graphql::Response,
-    SupergraphAttributes,
-    SupergraphSelector,
->;
-
-pub(crate) type SubgraphEvents =
-    CustomEvents<subgraph::Request, subgraph::Response, (), SubgraphAttributes, SubgraphSelector>;
-
 pub(crate) struct CustomEvents<Request, Response, EventResponse, Attributes, Sel>
 where
     Attributes: Selectors<Request, Response, EventResponse> + Default,
@@ -199,315 +174,6 @@ where
     pub(super) response: Option<StandardEvent<Sel>>,
     pub(super) error: Option<StandardEvent<Sel>>,
     pub(super) custom: Vec<CustomEvent<Request, Response, EventResponse, Attributes, Sel>>,
-}
-
-impl CustomEvents<router::Request, router::Response, (), RouterAttributes, RouterSelector> {
-    pub(crate) fn on_request(&mut self, request: &router::Request) {
-        if let Some(request_event) = &mut self.request {
-            if request_event.condition.evaluate_request(request) != Some(true) {
-                return;
-            }
-
-            request
-                .context
-                .extensions()
-                .with_lock(|ext| ext.insert(DisplayRouterRequest(request_event.level)));
-        }
-        if self.response.is_some() {
-            request
-                .context
-                .extensions()
-                .with_lock(|ext| ext.insert(DisplayRouterResponse(true)));
-        }
-        for custom_event in &mut self.custom {
-            custom_event.on_request(request);
-        }
-    }
-
-    pub(crate) fn on_response(&mut self, response: &router::Response) {
-        if let Some(response_event) = &self.response {
-            if !response_event.condition.evaluate_response(response) {
-                return;
-            }
-            let mut attrs = Vec::with_capacity(4);
-
-            #[cfg(test)]
-            let mut headers: indexmap::IndexMap<String, HeaderValue> = response
-                .response
-                .headers()
-                .clone()
-                .into_iter()
-                .filter_map(|(name, val)| Some((name?.to_string(), val)))
-                .collect();
-            #[cfg(test)]
-            headers.sort_keys();
-            #[cfg(not(test))]
-            let headers = response.response.headers();
-            attrs.push(KeyValue::new(
-                HTTP_RESPONSE_HEADERS,
-                opentelemetry::Value::String(format!("{:?}", headers).into()),
-            ));
-            attrs.push(KeyValue::new(
-                HTTP_RESPONSE_STATUS,
-                opentelemetry::Value::String(format!("{}", response.response.status()).into()),
-            ));
-            attrs.push(KeyValue::new(
-                HTTP_RESPONSE_VERSION,
-                opentelemetry::Value::String(format!("{:?}", response.response.version()).into()),
-            ));
-
-            if let Some(body) = response
-                .context
-                .extensions()
-                .with_lock(|ext| ext.remove::<RouterResponseBodyExtensionType>())
-            {
-                attrs.push(KeyValue::new(
-                    HTTP_RESPONSE_BODY,
-                    opentelemetry::Value::String(body.0.into()),
-                ));
-            }
-
-            log_event(response_event.level, "router.response", attrs, "");
-        }
-        for custom_event in &mut self.custom {
-            custom_event.on_response(response);
-        }
-    }
-
-    pub(crate) fn on_error(&mut self, error: &BoxError, ctx: &Context) {
-        if let Some(error_event) = &self.error {
-            if !error_event.condition.evaluate_error(error, ctx) {
-                return;
-            }
-            log_event(
-                error_event.level,
-                "router.error",
-                vec![KeyValue::new(
-                    Key::from_static_str("error"),
-                    opentelemetry::Value::String(error.to_string().into()),
-                )],
-                "",
-            );
-        }
-        for custom_event in &mut self.custom {
-            custom_event.on_error(error, ctx);
-        }
-    }
-}
-
-impl
-    CustomEvents<
-        supergraph::Request,
-        supergraph::Response,
-        graphql::Response,
-        SupergraphAttributes,
-        SupergraphSelector,
-    >
-{
-    pub(crate) fn on_request(&mut self, request: &supergraph::Request) {
-        if let Some(request_event) = &mut self.request {
-            if request_event.condition.evaluate_request(request) != Some(true) {
-                return;
-            }
-            let mut attrs = Vec::with_capacity(5);
-            #[cfg(test)]
-            let mut headers: indexmap::IndexMap<String, HeaderValue> = request
-                .supergraph_request
-                .headers()
-                .clone()
-                .into_iter()
-                .filter_map(|(name, val)| Some((name?.to_string(), val)))
-                .collect();
-            #[cfg(test)]
-            headers.sort_keys();
-            #[cfg(not(test))]
-            let headers = request.supergraph_request.headers();
-            attrs.push(KeyValue::new(
-                HTTP_REQUEST_HEADERS,
-                opentelemetry::Value::String(format!("{:?}", headers).into()),
-            ));
-            attrs.push(KeyValue::new(
-                HTTP_REQUEST_METHOD,
-                opentelemetry::Value::String(
-                    format!("{}", request.supergraph_request.method()).into(),
-                ),
-            ));
-            attrs.push(KeyValue::new(
-                HTTP_REQUEST_URI,
-                opentelemetry::Value::String(
-                    format!("{}", request.supergraph_request.uri()).into(),
-                ),
-            ));
-            attrs.push(KeyValue::new(
-                HTTP_REQUEST_VERSION,
-                opentelemetry::Value::String(
-                    format!("{:?}", request.supergraph_request.version()).into(),
-                ),
-            ));
-            attrs.push(KeyValue::new(
-                HTTP_REQUEST_BODY,
-                opentelemetry::Value::String(
-                    serde_json::to_string(request.supergraph_request.body())
-                        .unwrap_or_default()
-                        .into(),
-                ),
-            ));
-            log_event(request_event.level, "supergraph.request", attrs, "");
-        }
-        if let Some(response_event) = self.response.take() {
-            request.context.extensions().with_lock(|lock| {
-                lock.insert(SupergraphEventResponse {
-                    level: response_event.level,
-                    condition: Arc::new(response_event.condition),
-                })
-            });
-        }
-        for custom_event in &mut self.custom {
-            custom_event.on_request(request);
-        }
-    }
-
-    pub(crate) fn on_response(&mut self, response: &supergraph::Response) {
-        for custom_event in &mut self.custom {
-            custom_event.on_response(response);
-        }
-    }
-
-    pub(crate) fn on_response_event(&self, response: &graphql::Response, ctx: &Context) {
-        for custom_event in &self.custom {
-            custom_event.on_response_event(response, ctx);
-        }
-    }
-
-    pub(crate) fn on_error(&mut self, error: &BoxError, ctx: &Context) {
-        if let Some(error_event) = &self.error {
-            if !error_event.condition.evaluate_error(error, ctx) {
-                return;
-            }
-            log_event(
-                error_event.level,
-                "supergraph.error",
-                vec![KeyValue::new(
-                    Key::from_static_str("error"),
-                    opentelemetry::Value::String(error.to_string().into()),
-                )],
-                "",
-            );
-        }
-        for custom_event in &mut self.custom {
-            custom_event.on_error(error, ctx);
-        }
-    }
-}
-
-impl CustomEvents<subgraph::Request, subgraph::Response, (), SubgraphAttributes, SubgraphSelector> {
-    pub(crate) fn on_request(&mut self, request: &subgraph::Request) {
-        if let Some(mut request_event) = self.request.take() {
-            if request_event.condition.evaluate_request(request) != Some(true) {
-                return;
-            }
-            request.context.extensions().with_lock(|lock| {
-                lock.insert(SubgraphEventRequest {
-                    level: request_event.level,
-                    condition: Arc::new(Mutex::new(request_event.condition)),
-                })
-            });
-        }
-        if let Some(response_event) = self.response.take() {
-            request.context.extensions().with_lock(|lock| {
-                lock.insert(SubgraphEventResponse {
-                    level: response_event.level,
-                    condition: Arc::new(response_event.condition),
-                })
-            });
-        }
-        for custom_event in &mut self.custom {
-            custom_event.on_request(request);
-        }
-    }
-
-    pub(crate) fn on_response(&mut self, response: &subgraph::Response) {
-        for custom_event in &mut self.custom {
-            custom_event.on_response(response);
-        }
-    }
-
-    pub(crate) fn on_error(&mut self, error: &BoxError, ctx: &Context) {
-        if let Some(error_event) = &self.error {
-            if !error_event.condition.evaluate_error(error, ctx) {
-                return;
-            }
-            log_event(
-                error_event.level,
-                "subgraph.error",
-                vec![KeyValue::new(
-                    Key::from_static_str("error"),
-                    opentelemetry::Value::String(error.to_string().into()),
-                )],
-                "",
-            );
-        }
-        for custom_event in &mut self.custom {
-            custom_event.on_error(error, ctx);
-        }
-    }
-}
-
-#[derive(Clone, Deserialize, JsonSchema, Debug, Default)]
-#[serde(deny_unknown_fields, default)]
-struct RouterEventsConfig {
-    /// Log the router request
-    request: StandardEventConfig<RouterSelector>,
-    /// Log the router response
-    response: StandardEventConfig<RouterSelector>,
-    /// Log the router error
-    error: StandardEventConfig<RouterSelector>,
-}
-
-#[derive(Clone)]
-pub(crate) struct SupergraphEventResponse {
-    // XXX(@IvanGoncharov): As part of removing Arc from StandardEvent I moved it here
-    // I think it's not nessary here but can't verify it right now, so in future can just wrap StandardEvent
-    pub(crate) level: EventLevel,
-    pub(crate) condition: Arc<Condition<SupergraphSelector>>,
-}
-
-#[derive(Clone)]
-pub(crate) struct SubgraphEventResponse {
-    // XXX(@IvanGoncharov): As part of removing Arc from StandardEvent I moved it here
-    // I think it's not nessary here but can't verify it right now, so in future can just wrap StandardEvent
-    pub(crate) level: EventLevel,
-    pub(crate) condition: Arc<Condition<SubgraphSelector>>,
-}
-
-#[derive(Clone)]
-pub(crate) struct SubgraphEventRequest {
-    // XXX(@IvanGoncharov): As part of removing Mutex from StandardEvent I moved it here
-    // I think it's not nessary here but can't verify it right now, so in future can just wrap StandardEvent
-    pub(crate) level: EventLevel,
-    pub(crate) condition: Arc<Mutex<Condition<SubgraphSelector>>>,
-}
-
-#[derive(Clone, Deserialize, JsonSchema, Debug, Default)]
-#[serde(deny_unknown_fields, default)]
-struct SupergraphEventsConfig {
-    /// Log the supergraph request
-    request: StandardEventConfig<SupergraphSelector>,
-    /// Log the supergraph response
-    response: StandardEventConfig<SupergraphSelector>,
-    /// Log the supergraph error
-    error: StandardEventConfig<SupergraphSelector>,
-}
-
-#[derive(Clone, Deserialize, JsonSchema, Debug, Default)]
-#[serde(deny_unknown_fields, default)]
-struct SubgraphEventsConfig {
-    /// Log the subgraph request
-    request: StandardEventConfig<SubgraphSelector>,
-    /// Log the subgraph response
-    response: StandardEventConfig<SubgraphSelector>,
-    /// Log the subgraph error
-    error: StandardEventConfig<SubgraphSelector>,
 }
 
 #[derive(Deserialize, JsonSchema, Clone, Debug)]
@@ -759,540 +425,5 @@ pub(crate) fn log_event(level: EventLevel, kind: &str, attributes: Vec<KeyValue>
         EventLevel::Error => {
             ::tracing::error!(%kind, "{}", message)
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::str::FromStr;
-
-    use apollo_compiler::name;
-    use apollo_federation::sources::connect::ConnectId;
-    use apollo_federation::sources::connect::ConnectSpec;
-    use apollo_federation::sources::connect::Connector;
-    use apollo_federation::sources::connect::HttpJsonTransport;
-    use apollo_federation::sources::connect::JSONSelection;
-    use apollo_federation::sources::connect::StringTemplate;
-    use http::HeaderValue;
-    use http::header::CONTENT_LENGTH;
-    use router::body;
-    use tracing::instrument::WithSubscriber;
-
-    use super::*;
-    use crate::assert_snapshot_subscriber;
-    use crate::context::CONTAINS_GRAPHQL_ERROR;
-    use crate::context::OPERATION_NAME;
-    use crate::graphql;
-    use crate::plugins::connectors::handle_responses::MappedResponse;
-    use crate::plugins::connectors::make_requests::ResponseKey;
-    use crate::plugins::telemetry::Telemetry;
-    use crate::plugins::test::PluginTestHarness;
-    use crate::services::connector::request_service::Request;
-    use crate::services::connector::request_service::Response;
-    use crate::services::connector::request_service::TransportRequest;
-    use crate::services::connector::request_service::TransportResponse;
-    use crate::services::connector::request_service::transport;
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_router_events() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            test_harness
-                .router_service(
-                    |_r|async  {
-                        Ok(router::Response::fake_builder()
-                            .header("custom-header", "val1")
-                            .header(CONTENT_LENGTH, "25")
-                            .header("x-log-request", HeaderValue::from_static("log"))
-                            .data(serde_json_bytes::json!({"data": "res"}))
-                            .build()
-                            .expect("expecting valid response"))
-                    },
-                )
-                .call(
-                    router::Request::fake_builder()
-                        .header(CONTENT_LENGTH, "0")
-                        .header("custom-header", "val1")
-                        .header("x-log-request", HeaderValue::from_static("log"))
-                        .build()
-                        .unwrap()
-                )
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(
-            assert_snapshot_subscriber!({r#"[].span["apollo_private.duration_ns"]"# => "[duration]", r#"[].spans[]["apollo_private.duration_ns"]"# => "[duration]", "[].fields.attributes" => insta::sorted_redaction()}),
-        )
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_router_events_graphql_error() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            // Without the header to enable custom event
-            test_harness
-                .router_service(
-
-                    |_r| async {
-                        let context_with_error = Context::new();
-                        let _ = context_with_error
-                            .insert(CONTAINS_GRAPHQL_ERROR, true)
-                            .unwrap();
-                        Ok(router::Response::fake_builder()
-                            .header("custom-header", "val1")
-                            .context(context_with_error)
-                            .data(serde_json_bytes::json!({"errors": [{"message": "res"}]}))
-                            .build()
-                            .expect("expecting valid response"))
-                    },
-                )
-                .call(router::Request::fake_builder()
-                    .header("custom-header", "val1")
-                    .build()
-                    .unwrap())
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(
-            assert_snapshot_subscriber!({r#"[].span["apollo_private.duration_ns"]"# => "[duration]", r#"[].spans[]["apollo_private.duration_ns"]"# => "[duration]", "[].fields.attributes" => insta::sorted_redaction()}),
-        )
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_router_events_graphql_response() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            // Without the header to enable custom event
-            test_harness
-                .router_service(
-                    |_r| async {
-                        Ok(router::Response::fake_builder()
-                            .header("custom-header", "val1")
-                            .header(CONTENT_LENGTH, "25")
-                            .header("x-log-response", HeaderValue::from_static("log"))
-                            .data(serde_json_bytes::json!({"data": "res"}))
-                            .build()
-                            .expect("expecting valid response"))
-                    },
-                )
-                .call(router::Request::fake_builder()
-                    .header("custom-header", "val1")
-                    .build()
-                    .unwrap())
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(
-            assert_snapshot_subscriber!({r#"[].span["apollo_private.duration_ns"]"# => "[duration]", r#"[].spans[]["apollo_private.duration_ns"]"# => "[duration]", "[].fields.attributes" => insta::sorted_redaction()}),
-        )
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_supergraph_events() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            test_harness
-                .supergraph_service(|_r| async {
-                    supergraph::Response::fake_builder()
-                        .header("custom-header", "val1")
-                        .header("x-log-request", HeaderValue::from_static("log"))
-                        .data(serde_json::json!({"data": "res"}).to_string())
-                        .build()
-                })
-                .call(
-                    supergraph::Request::fake_builder()
-                        .query("query { foo }")
-                        .header("x-log-request", HeaderValue::from_static("log"))
-                        .build()
-                        .unwrap(),
-                )
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(assert_snapshot_subscriber!())
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_supergraph_events_with_exists_condition() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!(
-                "../testdata/custom_events_exists_condition.router.yaml"
-            ))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            let ctx = Context::new();
-            ctx.insert(OPERATION_NAME, String::from("Test")).unwrap();
-            test_harness
-                .supergraph_service(|_r| async {
-                    supergraph::Response::fake_builder()
-                        .data(serde_json::json!({"data": "res"}).to_string())
-                        .build()
-                })
-                .call(
-                    supergraph::Request::fake_builder()
-                        .query("query Test { foo }")
-                        .context(ctx)
-                        .build()
-                        .unwrap(),
-                )
-                .await
-                .expect("expecting successful response");
-            test_harness
-                .supergraph_service(|_r| async {
-                    Ok(supergraph::Response::fake_builder()
-                        .data(serde_json::json!({"data": "res"}).to_string())
-                        .build()
-                        .expect("expecting valid response"))
-                })
-                .call(
-                    supergraph::Request::fake_builder()
-                        .query("query { foo }")
-                        .build()
-                        .unwrap(),
-                )
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(assert_snapshot_subscriber!())
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_supergraph_events_on_graphql_error() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            test_harness
-                .supergraph_service(|_r| async {
-                    let context_with_error = Context::new();
-                    let _ = context_with_error
-                        .insert(CONTAINS_GRAPHQL_ERROR, true)
-                        .unwrap();
-                    supergraph::Response::fake_builder()
-                        .header("custom-header", "val1")
-                        .header("x-log-request", HeaderValue::from_static("log"))
-                        .context(context_with_error)
-                        .data(serde_json_bytes::json!({"errors": [{"message": "res"}]}))
-                        .build()
-                })
-                .call(
-                    supergraph::Request::fake_builder()
-                        .query("query { foo }")
-                        .build()
-                        .unwrap(),
-                )
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(assert_snapshot_subscriber!())
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_supergraph_events_on_response() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            test_harness
-                .supergraph_service(|_r| async {
-                    supergraph::Response::fake_builder()
-                        .header("custom-header", "val1")
-                        .header("x-log-response", HeaderValue::from_static("log"))
-                        .data(serde_json_bytes::json!({"errors": [{"message": "res"}]}))
-                        .build()
-                })
-                .call(
-                    supergraph::Request::fake_builder()
-                        .query("query { foo }")
-                        .build()
-                        .unwrap(),
-                )
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(assert_snapshot_subscriber!())
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_subgraph_events() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            let mut subgraph_req = http::Request::new(
-                graphql::Request::fake_builder()
-                    .query("query { foo }")
-                    .build(),
-            );
-            subgraph_req
-                .headers_mut()
-                .insert("x-log-request", HeaderValue::from_static("log"));
-            test_harness
-                .subgraph_service("subgraph", |_r| async {
-                    subgraph::Response::fake2_builder()
-                        .header("custom-header", "val1")
-                        .header("x-log-request", HeaderValue::from_static("log"))
-                        .data(serde_json::json!({"data": "res"}).to_string())
-                        .build()
-                })
-                .call(
-                    subgraph::Request::fake_builder()
-                        .subgraph_name("subgraph")
-                        .subgraph_request(subgraph_req)
-                        .build(),
-                )
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(assert_snapshot_subscriber!())
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_subgraph_events_response() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            let mut subgraph_req = http::Request::new(
-                graphql::Request::fake_builder()
-                    .query("query { foo }")
-                    .build(),
-            );
-            subgraph_req
-                .headers_mut()
-                .insert("x-log-request", HeaderValue::from_static("log"));
-            test_harness
-                .subgraph_service("subgraph", |_r| async {
-                    subgraph::Response::fake2_builder()
-                        .header("custom-header", "val1")
-                        .header("x-log-response", HeaderValue::from_static("log"))
-                        .subgraph_name("subgraph")
-                        .data(serde_json::json!({"data": "res"}).to_string())
-                        .build()
-                })
-                .call(
-                    subgraph::Request::fake_builder()
-                        .subgraph_name("subgraph")
-                        .subgraph_request(subgraph_req)
-                        .build(),
-                )
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(assert_snapshot_subscriber!())
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_connector_events_request() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            let context = crate::Context::default();
-            let mut http_request = http::Request::builder().body("".into()).unwrap();
-            http_request
-                .headers_mut()
-                .insert("x-log-request", HeaderValue::from_static("log"));
-            let transport_request = TransportRequest::Http(transport::http::HttpRequest {
-                inner: http_request,
-                debug: None,
-            });
-            let connector = Connector {
-                id: ConnectId::new(
-                    "subgraph".into(),
-                    Some("source".into()),
-                    name!(Query),
-                    name!(users),
-                    0,
-                    "label",
-                ),
-                transport: HttpJsonTransport {
-                    source_url: None,
-                    connect_template: StringTemplate::from_str("/test").unwrap(),
-                    ..Default::default()
-                },
-                selection: JSONSelection::empty(),
-                config: None,
-                max_requests: None,
-                entity_resolver: None,
-                spec: ConnectSpec::V0_1,
-                request_variables: Default::default(),
-                response_variables: Default::default(),
-                batch_settings: None,
-                request_headers: Default::default(),
-                response_headers: Default::default(),
-            };
-            let response_key = ResponseKey::RootField {
-                name: "hello".to_string(),
-                inputs: Default::default(),
-                selection: Arc::new(JSONSelection::parse("$.data").unwrap()),
-            };
-            let connector_request = Request {
-                context: context.clone(),
-                connector: Arc::new(connector.clone()),
-                service_name: Default::default(),
-                transport_request,
-                key: response_key.clone(),
-                mapping_problems: vec![],
-                supergraph_request: Default::default(),
-            };
-            test_harness
-                .call_connector_request_service(connector_request, |request| Response {
-                    context: request.context.clone(),
-                    connector: request.connector.clone(),
-                    transport_result: Ok(TransportResponse::Http(transport::http::HttpResponse {
-                        inner: http::Response::builder()
-                            .status(200)
-                            .header("x-log-request", HeaderValue::from_static("log"))
-                            .body(body::empty())
-                            .expect("expecting valid response")
-                            .into_parts()
-                            .0,
-                    })),
-                    mapped_response: MappedResponse::Data {
-                        data: serde_json::json!({})
-                            .try_into()
-                            .expect("expecting valid JSON"),
-                        key: request.key.clone(),
-                        problems: vec![],
-                    },
-                })
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(assert_snapshot_subscriber!())
-        .await
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_connector_events_response() {
-        let test_harness: PluginTestHarness<Telemetry> = PluginTestHarness::builder()
-            .config(include_str!("../testdata/custom_events.router.yaml"))
-            .build()
-            .await
-            .expect("test harness");
-
-        async {
-            let context = crate::Context::default();
-            let mut http_request = http::Request::builder().body("".into()).unwrap();
-            http_request
-                .headers_mut()
-                .insert("x-log-response", HeaderValue::from_static("log"));
-            let transport_request = TransportRequest::Http(transport::http::HttpRequest {
-                inner: http_request,
-                debug: None,
-            });
-            let connector = Connector {
-                id: ConnectId::new(
-                    "subgraph".into(),
-                    Some("source".into()),
-                    name!(Query),
-                    name!(users),
-                    0,
-                    "label",
-                ),
-                transport: HttpJsonTransport {
-                    source_url: None,
-                    connect_template: StringTemplate::from_str("/test").unwrap(),
-                    ..Default::default()
-                },
-                selection: JSONSelection::empty(),
-                config: None,
-                max_requests: None,
-                entity_resolver: None,
-                spec: ConnectSpec::V0_1,
-                request_variables: Default::default(),
-                response_variables: Default::default(),
-                batch_settings: None,
-                request_headers: Default::default(),
-                response_headers: Default::default(),
-            };
-            let response_key = ResponseKey::RootField {
-                name: "hello".to_string(),
-                inputs: Default::default(),
-                selection: Arc::new(JSONSelection::parse("$.data").unwrap()),
-            };
-            let connector_request = Request {
-                context: context.clone(),
-                connector: Arc::new(connector.clone()),
-                service_name: Default::default(),
-                transport_request,
-                key: response_key.clone(),
-                mapping_problems: vec![],
-                supergraph_request: Default::default(),
-            };
-            test_harness
-                .call_connector_request_service(connector_request, |request| Response {
-                    context: request.context.clone(),
-                    connector: request.connector.clone(),
-                    transport_result: Ok(TransportResponse::Http(transport::http::HttpResponse {
-                        inner: http::Response::builder()
-                            .status(200)
-                            .header("x-log-response", HeaderValue::from_static("log"))
-                            .body(body::empty())
-                            .expect("expecting valid response")
-                            .into_parts()
-                            .0,
-                    })),
-                    mapped_response: MappedResponse::Data {
-                        data: serde_json::json!({})
-                            .try_into()
-                            .expect("expecting valid JSON"),
-                        key: request.key.clone(),
-                        problems: vec![],
-                    },
-                })
-                .await
-                .expect("expecting successful response");
-        }
-        .with_subscriber(assert_snapshot_subscriber!())
-        .await
     }
 }
