+++ apollo-router/src/services/subgraph_service.rs
@@ -68,9 +68,9 @@ use crate::plugins::subscription::SubscriptionConfig;
 use crate::plugins::subscription::SubscriptionMode;
 use crate::plugins::subscription::WebSocketConfiguration;
 use crate::plugins::subscription::create_verifier;
-use crate::plugins::telemetry::config_new::events::SubgraphEventRequest;
-use crate::plugins::telemetry::config_new::events::SubgraphEventResponse;
 use crate::plugins::telemetry::config_new::events::log_event;
+use crate::plugins::telemetry::config_new::subgraph::events::SubgraphEventRequest;
+use crate::plugins::telemetry::config_new::subgraph::events::SubgraphEventResponse;
 use crate::plugins::telemetry::consts::SUBGRAPH_REQUEST_SPAN_NAME;
 use crate::protocols::websocket::GraphqlWebSocket;
 use crate::protocols::websocket::convert_websocket_stream;
